Update NewGRF URLs after BaNaNaS registry update
BaNaNaS no longer serves NewGRF under user friendly URLs and pushes
towards downloading packages with the in-game BaNaNas client.

This quick fix changes download-newgrf.sh to use the new URLs as
returned from the API. Finding these URLs is not trivial, the way that
NewGRFs are configured in this project should be changed.